requirements.txt
python-telegram-bot==21.6
yt-dlp>=2024.11.4
rubpy>=6.9.0
aiohttp>=3.9.0
